% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIVDGP.R
\name{BIVDGP}
\alias{BIVDGP}
\title{Normal errors with break in variance}
\usage{
BIVDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{

}
\examples{
BIVDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
