% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauchyDGP.R
\name{CauchyDGP}
\alias{CauchyDGP}
\title{Cauchy perturbations}
\usage{
CauchyDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim Cauchy}. An example of a predictive regression
DGP with Cauchy perturbations can be found in \insertCite{dufour1995exact;textual}{Rdpack}.
}
\examples{
CauchyDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
