% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpVarDGP.R
\name{ExpVarDGP}
\alias{ExpVarDGP}
\title{Normal errors with exponential variance}
\usage{
ExpVarDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim N(0,\sigma_t^2)},
where \eqn{\sigma_t=\exp(0.5t)}. Examples of DGPs with Normal disturbances and exponential variance can be found in
\insertCite{dufour2010exact;textual}{Rdpack} and \insertCite{coudin2009finite;textual}{Rdpack}.
}
\examples{
ExpVarDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
