% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarchDGP.R
\name{GarchDGP}
\alias{GarchDGP}
\title{Normal errors with stationary GARCH(1,1) variance}
\usage{
GarchDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim N(0,\sigma_t^2)},
where \eqn{\sigma_t^2=0.00037+0.0888\varepsilon_{t-1}^2+0.9024\sigma_{t-1}^2}. Examples of DGPs with Normal disturbances and stationary GARCH(1,1) variance can be found
in \insertCite{dufour2010exact;textual}{Rdpack} and \insertCite{coudin2009finite;textual}{Rdpack}.
}
\examples{
GarchDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
