% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureDGP.R
\name{MixtureDGP}
\alias{MixtureDGP}
\title{Mixture of Cauchy and Normal errors}
\usage{
MixtureDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim s_t\lvert\varepsilon_t^C\rvert-(1-s_t)\lvert\varepsilon_t^N\rvert},
where \eqn{P(s_t=0)=P(s_t=1)=0.5} for all \eqn{t}. An example of a DGP with mixture perturbations can be found in \insertCite{dufour2010exact;textual}{Rdpack}.
}
\examples{
MixtureDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
