% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalDGP.R
\name{NormalDGP}
\alias{NormalDGP}
\title{Standard Normal errors}
\usage{
NormalDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The original framework initially studied by \insertCite{mankiw1986we;textual}{Rdpack}.
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim N(0,1)}.
}
\examples{
NormalDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
