% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentDGP.R
\name{StudentDGP}
\alias{StudentDGP}
\title{Student's \eqn{t(2)} errors}
\usage{
StudentDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations.}

\item{beta}{the regressor coefficient of the predictive regression.}

\item{theta}{the autocorrelation coefficient of the predictor.}

\item{rho}{the contemporaneous correlation coefficient.}
}
\description{
The errors of the predictive regression are distributed according to \eqn{\varepsilon_t\sim t(2)}. An example of a predictive regression
DGP with \eqn{t(3)} perturbations can be found in \insertCite{dufour1995exact;textual}{Rdpack}. Outside a predictive regression framework, \eqn{t(2)} disturbances have further been considered in
\insertCite{dufour2010exact;textual}{Rdpack}.
}
\examples{
StudentDGP(50,0.5,0.999,0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity,}
\keyword{Normal}
\keyword{persistency,}
