% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureDGP.R
\name{MixtureDGP}
\alias{MixtureDGP}
\title{Mixture of Cauchy and Normal errors}
\usage{
MixtureDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations}

\item{beta}{the regressor coefficient of the predictive regression}

\item{theta}{the autocorrelation coefficient of the predictor}

\item{rho}{the contemporaneous correlation coefficient}
}
\description{
The processes \eqn{y} and \eqn{x} are generated by a predictive regression of the form \eqn{y_t=\beta+x_{t-1}+\varepsilon_t} for \eqn{t=1,\cdots,T},
in which the regressors follow an AR(1) process - i.e. \eqn{x_t=\theta x_{t-1}+u_t}.
The predictor's errors are distributed according to \eqn{u_t\sim N(0,1)}, whereas the disturbances
of the predictive regression, \eqn{\varepsilon_t}, are distributed \eqn{\varepsilon_t\sim s_t|\varepsilon_t^C|-(1-s_t)|\varepsilon_t^N|},
where \eqn{P(s_t=0)=P(s_t=1)=0.5} for all \eqn{t}. An example of a DGP with mixture perturbations can be found in \insertCite{dufour2010exact;textual}{PredictiveDGP}. The initial value of the process \eqn{x} is generated by \eqn{x_0=\frac{w_0}{\sqrt{1-\theta^2}}},
where \eqn{w_t\sim N(0,1)}. Finally, the contemporaneous correlation between the disturbances \eqn{\varepsilon_t} and \eqn{u_t} is captured by \eqn{\rho\varepsilon_t+w_t\sqrt{1-\rho^2}}.
}
\examples{
MixtureDGP(n=50, beta=0.5, theta=0.999, rho=0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity}
\keyword{mixture}
\keyword{normal}
\keyword{persistency}
