% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentDGP.R
\name{StudentDGP}
\alias{StudentDGP}
\title{Student's \eqn{t(2)} errors}
\usage{
StudentDGP(n, beta, theta, rho)
}
\arguments{
\item{n}{the number of observations}

\item{beta}{the regressor coefficient of the predictive regression}

\item{theta}{the autocorrelation coefficient of the predictor}

\item{rho}{the contemporaneous correlation coefficient}
}
\description{
The processes \eqn{y} and \eqn{x} are generated by a predictive regression of the form \eqn{y_t=\beta+x_{t-1}+\varepsilon_t} for \eqn{t=1,\cdots,T},
in which the regressors follow an AR(1) process - i.e. \eqn{x_t=\theta x_{t-1}+u_t}.
The predictor's errors are distributed according to \eqn{u_t\sim N(0,1)}, whereas the disturbances
of the predictive regression, \eqn{\varepsilon_t}, are distributed \eqn{\varepsilon_t\sim t(2)}. An example of a predictive regression
DGP with \eqn{t(3)} perturbations can be found in \insertCite{dufour1995exact;textual}{PredictiveDGP}. Outside a predictive regression framework, \eqn{t(2)} disturbances have further been considered in
\insertCite{dufour2010exact;textual}{PredictiveDGP}. The initial value of the process \eqn{x} is generated by \eqn{x_0=\frac{w_0}{\sqrt{1-\theta^2}}},
where \eqn{w_t\sim N(0,1)}. Finally, the contemporaneous correlation between the disturbances \eqn{\varepsilon_t} and \eqn{u_t} is captured by \eqn{\rho\varepsilon_t+w_t\sqrt{1-\rho^2}}.
}
\examples{
StudentDGP(n=50, p=0.5, rho=0.999, theta=0.9)
}
\references{
\insertAllCited{}
}
\keyword{Endogeneity}
\keyword{Student-t}
\keyword{persistency}
